requirements.txt
asn1crypto==1.1.0
backports.shutil-get-terminal-size==1.0.0
certifi==2019.9.11
cffi==1.12.3
chardet==3.0.4
cryptography==2.7
decorator==4.0.10
Django==2.2.1
django-crispy-forms==1.8.1
djangorestframework==3.10.2
enum34==1.1.6
html5lib==1.0.1
idna==2.8
ipython-genutils==0.1.0
Markdown==3.1.1
pathlib2==2.1.0
pexpect==4.2.1
pickleshare==0.7.4
Pillow==6.1.0
prompt-toolkit==1.0.9
ptyprocess==0.5.1
pycryptodome==3.9.0
Pygments==2.1.3
PyJWT==1.7.1
PyPDF2==1.26.0
PySocks==1.7.0
pytz==2019.1
reportlab==3.5.23
requests==2.22.0
simplegeneric==0.8.1
six==1.12.0
sqlparse==0.3.0
traitlets==4.3.1
twilio==6.29.2
urllib3==1.25.3
wcwidth==0.1.7
webencodings==0.5.1
wincertstore==0.2
xhtml2pdf==0.2.3
